chore(): clean up and add tests for all 3 currently supported stream libraries
This relies on a fix to xstream that I have locally and will PR shortly.
