--- .eslintrc
@@ -1,11 +1,0 @@
-{
-  "extends": "eslint-config-cycle",
-  "env": {
-    "browser": true,
-    "node": true
-  },
-  "parser": "babel-eslint",
-  "rules": {
-    "quotes": 0,
-  }
-}
