+++ package.json
@@ -18,13 +18,15 @@
   },
   "homepage": "https://github.com/TylorS/cyclic-router#readme",
   "dependencies": {
-    "history": "^2.1.0"
+    "@cycle/history": "^3.0.4",
+    "history": "^2.1.0",
+    "switch-path": "^1.1.6"
   },
   "devDependencies": {
     "@cycle/base": "^2.0.1",
-    "@cycle/dom": "^10.0.0-rc8",
+    "@cycle/rx-adapter": "^1.0.3",
+    "@cycle/rxjs-adapter": "^1.0.5",
     "@cycle/xstream-adapter": "^1.0.2",
-    "@cycle/xstream-run": "^1.0.0",
     "assert": "^1.3.0",
     "babel-preset-es2015": "^6.6.0",
     "babel-register": "^6.7.2",
@@ -35,6 +37,7 @@
     "mkdirp": "^0.5.1",
     "mocha": "^2.4.5",
     "rx": "^4.1.0",
+    "rxjs": "^5.0.0-beta.7",
     "testem": "^1.6.0",
     "ts-node": "^0.7.2",
     "tslint": "^3.8.0",
