+++ typings.json
@@ -1,13 +1,7 @@
 {
   "name": "cyclic-router",
   "dependencies": {
-    "@cycle/history": "npm:@cycle/history",
-    "@cycle/xstream-adapter": "npm:@cycle/xstream-adapter",
-    "xstream": "npm:xstream"
-  },
-  "devDependencies": {
-    "@cycle/base": "npm:@cycle/base",
-    "@cycle/xstream-run": "npm:@cycle/xstream-run"
+    "@cycle/history": "npm:@cycle/history"
   },
   "ambientDevDependencies": {
     "mocha": "github:DefinitelyTyped/DefinitelyTyped/mocha/mocha.d.ts#d6dd320291705694ba8e1a79497a908e9f5e6617",
